Dialectic and phenomenon aspects
Dialectic angles and aspects added by this
commit help to understand what is agile and
what is not. What we want to achieve and what
we should avoid doing in agent smith.

They are the ethical representation of what is
agent smith. Agent smith should not be an agent
of control. Agent smith should not encourage
rigidity.

Individual is at the heart of agile, and fortunately
for us science comes with lot of identified
phenomenons that we can use to determine the
mission of agent smith.

There is a deep relation between the ethical aspects
and the phenomenon that we have handy. By using
them as a framework we can define clear goals
and above all we can define how to achieve them.